refactor: LottoMachine 메서드 역할 분리
