+++ src/main/java/lotto/model/LottoMachine.java
@@ -3,7 +3,6 @@ package lotto.model;
 import lotto.model.vo.LottoBall;
 
 import java.util.List;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -12,10 +11,14 @@ public class LottoMachine {
     private static final int MIN_LOTTO_NUM = 1;
     private static final int MAX_LOTTO_NUM = 45;
 
-    private static final List<LottoBall> LOTTO_NUMBERS = IntStream.rangeClosed(MIN_LOTTO_NUM, MAX_LOTTO_NUM)
-            .boxed()
-            .map(LottoBall::new)
-            .collect(Collectors.toList());
+    private static final List<LottoBall> LOTTO_NUMBERS = initializeLottoBalls();
+
+    private static List<LottoBall> initializeLottoBalls() {
+        return IntStream.rangeClosed(MIN_LOTTO_NUM, MAX_LOTTO_NUM)
+                .boxed()
+                .map(LottoBall::new)
+                .collect(Collectors.toList());
+    }
 
     private final LottoNumberSelector lottoNumberSelector;
 
@@ -24,18 +27,10 @@ public class LottoMachine {
     }
 
     public LottoGroup randomGenerate() {
-        return lottoNumberSelector.generate(LOTTO_NUMBERS);
+        return lottoNumberSelector.randomSelect(LOTTO_NUMBERS);
     }
 
     public LottoGroup manualGenerate(final List<Integer> numbers) {
-        return new LottoGroup(LOTTO_NUMBERS.stream()
-                .filter(getLottoBallPredicate(numbers))
-                .collect(Collectors.toUnmodifiableList())
-        );
-    }
-
-    private Predicate<LottoBall> getLottoBallPredicate(final List<Integer> numbers) {
-        return lottoBall -> numbers.stream()
-                .anyMatch(lottoBall::hasNumber);
+        return lottoNumberSelector.customSelect(LOTTO_NUMBERS, numbers);
     }
 }
