+++ src/main/java/lotto/model/LottoNumberSelector.java
@@ -4,8 +4,9 @@ import lotto.model.vo.LottoBall;
 
 import java.util.List;
 
-@FunctionalInterface
 public interface LottoNumberSelector {
 
-    LottoGroup generate(final List<LottoBall> lottoBalls);
+    LottoGroup randomSelect(final List<LottoBall> lottoBalls);
+
+    LottoGroup customSelect(final List<LottoBall> lottoBalls, final List<Integer>numbers);
 }
