--- src/main/java/lotto/model/RandomLottoNumberSelector.java
@@ -1,18 +1,0 @@
-package lotto.model;
-
-import lotto.model.vo.LottoBall;
-
-import java.util.Collections;
-import java.util.List;
-
-public class RandomLottoNumberSelector implements LottoNumberSelector {
-
-    private static final int ZERO = 0;
-    private static final int LOTTO_BALLS_COUNT = 6;
-
-    @Override
-    public LottoGroup generate(final List<LottoBall> lottoBalls) {
-        Collections.shuffle(lottoBalls);
-        return new LottoGroup(lottoBalls.subList(ZERO, LOTTO_BALLS_COUNT));
-    }
-}
