+++ src/test/java/lotto/model/TestLottoNumberGenerator.java
@@ -5,6 +5,7 @@ import lotto.model.vo.LottoBall;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 public class TestLottoNumberGenerator implements LottoNumberSelector {
@@ -22,7 +23,20 @@ public class TestLottoNumberGenerator implements LottoNumberSelector {
     }
 
     @Override
-    public LottoGroup generate(final List<LottoBall> lottoBalls) {
+    public LottoGroup randomSelect(final List<LottoBall> lottoBalls) {
         return lottoGroups.remove();
+    }
+
+    @Override
+    public LottoGroup customSelect(final List<LottoBall> lottoBalls, final List<Integer> numbers) {
+        return new LottoGroup(lottoBalls.stream()
+                .filter(getLottoBallPredicate(numbers))
+                .collect(Collectors.toUnmodifiableList())
+        );
+    }
+
+    private Predicate<LottoBall> getLottoBallPredicate(final List<Integer> numbers) {
+        return lottoBall -> numbers.stream()
+                .anyMatch(lottoBall::hasNumber);
     }
 }
